Make binary-safe
This commit makes the code binary-safe by replacing Scanner.Scan (which eats
the EOL characters) with ReadString (which preserves them).  For now the header
lines continue to have forced CRLF termination, but the body is unmodified so
that binaries that contain 0x0a don't get mangled.  The body data structure was
left as a slice of lines so we can still do CRLF modding if we want.

Unrelated changes:

* Do a better job of passing back errors on write.

* Small refactor to move the body into its own type.